requirements.txt
Jinja2==2.8
Markdown==2.6.7
MarkupSafe==0.23
pkg-resources==0.0.0
Pygments==2.1.3
python-slugify==1.2.1
Unidecode==0.4.19
